refactor: BonusNumber 클래스 삭제 -> LottoNumber 클래스로 기능 이동
